refactored AvailabilityMatrix.java
